Time

Documentation

data ClockTime

A representation of the internal clock time. Clock times may be compared, converted to strings, or converted to an external calendar time CalendarTime for I/O or other manipulations.

data Month

A month of the year.

Constructors

January

February

March

April

May

June

July

August

September

October

November

December

data Day

A day of the week.

Constructors

Sunday

Monday

Tuesday

Wednesday

Thursday

Friday

Saturday

data CalendarTime

CalendarTime is a user-readable and manipulable representation of the internal ClockTime type.

Constructors

CalendarTime

Fields

ctYear :: Int

Year (pre-Gregorian dates are inaccurate)

ctMonth :: Month

Month of the year

ctDay :: Int

Day of the month (1 to 31)

ctHour :: Int

Hour of the day (0 to 23)

ctMin :: Int

Minutes (0 to 59)

ctSec :: Int

Seconds (0 to 61, allowing for up to two leap seconds)

ctPicosec :: Integer

Picoseconds

ctWDay :: Day

Day of the week

ctYDay :: Int

Day of the year (0 to 364, or 365 in leap years)

ctTZName :: String

Name of the time zone

ctTZ :: Int

Variation from UTC in seconds

ctIsDST :: Bool

True if Daylight Savings Time would be in effect, and False otherwise

data TimeDiff

records the difference between two clock times in a user-readable way.

Constructors

TimeDiff

Fields

tdYear :: Int

tdMonth :: Int

tdDay :: Int

tdHour :: Int

tdMin :: Int

tdSec :: Int

tdPicosec :: Integer

getClockTime :: IO ClockTime

addToClockTime :: TimeDiff -> ClockTime -> ClockTime

addToClockTime d t adds a time difference d and a clock time t to yield a new clock time. The difference d may be either positive or negative.

diffClockTimes :: ClockTime -> ClockTime -> TimeDiff

diffClockTimes t1 t2 returns the difference between two clock times t1 and t2 as a TimeDiff.

toCalendarTime :: ClockTime -> IO CalendarTime

converts an internal clock time to a local time, modified by the timezone and daylight savings time settings in force at the time of conversion. Because of this dependence on the local environment, toCalendarTime is in the IO monad.

toUTCTime :: ClockTime -> CalendarTime

converts an internal clock time into a CalendarTime in standard UTC format.

toClockTime :: CalendarTime -> ClockTime

converts a CalendarTime into the corresponding internal ClockTime, ignoring the contents of the ctWDay, ctYDay, ctTZName and ctIsDST fields.

calendarTimeToString :: CalendarTime -> String

formats calendar times using local conventions.

formatCalendarTime :: TimeLocale -> String -> CalendarTime -> String

formats calendar times using local conventions and a formatting string. The formatting string is that understood by the ISO C strftime() function.